Add unit tests and fix distribution reliability
- Add ItemPoolFactory with all 10 case items
- Add BlockCalculatorTests (11 tests): block ranges, gaps, edge cases
- Add DistributionTests (9 tests): validity, randomness, impossible configs
- Fix float precision bug in BlockCalculator (float → double)
- Add retry mechanism to CardDistributor (max 100k iterations, 10 attempts)
- Set up EditMode test assembly with proper references

+++ Assets/Scripts/Core/Distribution/BlockCalculator.cs
@@ -16,7 +16,7 @@ namespace CardSelectionSystem.Core.Distribution
                 throw new ArgumentException("Item count cannot exceed total positions.");
 
             var blocks = new List<BlockRange>(itemCount);
-            float blockSize = (float)totalPositions / itemCount;
+            double blockSize = (double)totalPositions / itemCount;
 
             for (int i = 0; i < itemCount; i++)
             {
